+++ inc/push_swap.h
@@ -14,6 +14,7 @@
 # define PUSH_SWAP_H
 
 # include "stack.h"
+# include <stdio.h>
 
 typedef char	t_action;
 
@@ -37,8 +38,8 @@ typedef struct s_node
 	struct s_node	**adjacent_nodes;
 	struct s_node	*prev_node;
 	t_action		last_action;
-	int				cost;
-	int				heuristic;
+	double			cost;
+	double			heuristic;
 }	t_node;
 
 void		*parse_input(t_stack *stack_a, char **argv);
